requirements.txt
torch
transformers
sentence-transformers
scikit-learn
pandas
numpy
fastapi
uvicorn
pylint
datasets
accelerate
pydantic
mlflow
python-dotenv
pandas
kagglehub
torchvision
boto3
pyinstaller
nuitka
evaluate
rouge_score